mycasa-pro/requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
plotly>=5.18.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
requests>=2.31.0
python-dateutil>=2.8.0
pyyaml>=6.0
aiohttp>=3.9.0
yfinance>=0.2.36

# FastAPI backend
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# Document processing for SecondBrain uploads
pdfplumber>=0.10.0
python-docx>=1.1.0

# Database migrations
alembic>=1.13.0

# Async fix (allows nested event loops)
nest_asyncio>=1.6.0

# Semantic search embeddings
sentence-transformers>=2.2.0
anthropic
python-dotenv
openai>=1.12.0
pyjwt>=2.8.0
bcrypt>=4.1.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.0  # For async API testing
